backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sse-starlette==1.6.1
google-genai==0.2.0
firebase-admin==6.2.0
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.1
google-search-results==2.4.2
